+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/extensions/ChannelExtension.kt
@@ -17,6 +17,8 @@
 package io.getstream.chat.android.client.extensions
 
 import io.getstream.chat.android.client.ChatClient
+import io.getstream.chat.android.client.extensions.internal.containsUserMention
+import io.getstream.chat.android.client.extensions.internal.wasCreatedAfter
 import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.core.internal.InternalStreamChatApi
@@ -48,4 +50,25 @@ public fun Channel.getUsersExcludingCurrent(
     } else {
         users
     }
+}
+
+/**
+ * Counts messages in which [user] is mentioned.
+ * The method relies on the [Channel.messages] list and doesn't do any API call.
+ * Therefore, the count might be not reliable as it relies on the local data.
+ *
+ * @param user The User object for which unread mentions are counted.
+ *
+ * @return Number of messages containing unread user mention.
+ */
+public fun Channel.countUnreadMentionsForUser(user: User): Int {
+    val lastMessageSeenDate = read.firstOrNull { read -> read.user.id == user.id }?.lastRead
+
+    val messagesToCheck = if (lastMessageSeenDate == null) {
+        messages
+    } else {
+        messages.filter { message -> message.wasCreatedAfter(lastMessageSeenDate) }
+    }
+
+    return messagesToCheck.count { message -> message.containsUserMention(user) }
 }